access features only after login

+++ src/utility/playlist.utility.js
@@ -1,44 +1,52 @@
 import axios from "axios";
 
 export const axiosAddNewPlaylist = (newPlaylist, playlistDispatch, auth) => {
-  (async function () {
-    console.log(newPlaylist);
-    try {
-      const response = await axios.post(
-        "https://Vid-Lib-API-Forked.sayuk.repl.co/playlist",
+  if (auth) {
+    (async function () {
+      console.log(newPlaylist);
+      try {
+        const response = await axios.post(
+          "https://Vid-Lib-API-Forked.sayuk.repl.co/playlist",
 
-        { name: newPlaylist, videos: [] },
-        {
-          headers: {
-            "auth-token": auth.token,
-          },
-        }
-      );
+          { name: newPlaylist, videos: [] },
+          {
+            headers: {
+              "auth-token": auth.token,
+            },
+          }
+        );
 
-      playlistDispatch({ type: "CREATE_PLAYLIST", payload: response.data });
-    } catch (error) {
-      console.log(error);
-    }
-  })();
+        playlistDispatch({ type: "CREATE_PLAYLIST", payload: response.data });
+      } catch (error) {
+        console.log(error);
+      }
+    })();
+  } else {
+    console.log("LOOOGIN");
+  }
 };
 
 export function axiosRemovePlaylist(playlist, playlistDispatch, auth) {
-  (async function () {
-    try {
-      const response = await axios.delete(
-        `https://Vid-Lib-API-Forked.sayuk.repl.co/playlist/${playlist._id}`,
-        {
-          headers: {
-            "auth-token": auth.token,
-          },
-        }
-      );
+  if (auth) {
+    (async function () {
+      try {
+        const response = await axios.delete(
+          `https://Vid-Lib-API-Forked.sayuk.repl.co/playlist/${playlist._id}`,
+          {
+            headers: {
+              "auth-token": auth.token,
+            },
+          }
+        );
 
-      playlistDispatch({ type: "SET_PLAYLIST", payload: response.data });
-    } catch (error) {
-      console.log(error);
-    }
-  })();
+        playlistDispatch({ type: "SET_PLAYLIST", payload: response.data });
+      } catch (error) {
+        console.log(error);
+      }
+    })();
+  } else {
+    console.log("LOOOGIN");
+  }
 }
 
 export function axiosAddVideoToPlaylist(
@@ -47,44 +55,52 @@ export function axiosAddVideoToPlaylist(
   playlistDispatch,
   auth
 ) {
-  (async function () {
-    try {
-      const response = await axios.post(
-        `https://Vid-Lib-API-Forked.sayuk.repl.co/playlist/update/${playlist._id}`,
-        {
-          video: video,
-        },
-        {
-          headers: {
-            "auth-token": auth.token,
+  if (auth) {
+    (async function () {
+      try {
+        const response = await axios.post(
+          `https://Vid-Lib-API-Forked.sayuk.repl.co/playlist/update/${playlist._id}`,
+          {
+            video: video,
           },
-        }
-      );
-      console.log(video, playlist._id);
-      // playlistDispatch({ type: "SET_PLAYLIST", payload: response.data });
-    } catch (err) {
-      console.log(err);
-    }
-  })();
+          {
+            headers: {
+              "auth-token": auth.token,
+            },
+          }
+        );
+        console.log(video, playlist._id);
+        // playlistDispatch({ type: "SET_PLAYLIST", payload: response.data });
+      } catch (err) {
+        console.log(err);
+      }
+    })();
+  } else {
+    console.log("LOOOGIN");
+  }
 }
 
 export function axiosDeleteVideoFromPlaylist(playlist, video, auth) {
-  (async function () {
-    try {
-      const response = await axios.delete(
-        "https://Vid-Lib-API-Forked.sayuk.repl.co/playlist/delete",
-        {
-          playlistId: playlist,
-          videoId: video,
-        },
-        {
-          headers: {
-            "auth-token": auth.token,
+  if (auth) {
+    (async function () {
+      try {
+        const response = await axios.delete(
+          "https://Vid-Lib-API-Forked.sayuk.repl.co/playlist/delete",
+          {
+            playlistId: playlist,
+            videoId: video,
           },
-        }
-      );
-    } catch (err) {
-      console.log(err);
-    }
-  })();
+          {
+            headers: {
+              "auth-token": auth.token,
+            },
+          }
+        );
+      } catch (err) {
+        console.log(err);
+      }
+    })();
+  } else {
+    console.log("loooogin");
+  }
 }
